Restructure auth folder
